refactor: back references for getCurrentUser util functions
